优化基础服务类结构

+++ OpenAuth.App/Base/BaseIntAutoGenApp.cs
@@ -28,12 +28,12 @@ namespace OpenAuth.App
         /// 按id批量删除
         /// </summary>
         /// <param name="ids"></param>
-        public void Delete(long[] ids)
+        public void Delete(int[] ids)
         {
             Repository.Delete(u => ids.Contains(u.Id));
         }
 
-        public T Get(long id)
+        public T Get(int id)
         {
             return Repository.FirstOrDefault(u => u.Id == id);
         }
